Revision, Added Section (Case Classes)
The last class went a lot faster than expected. There were a lot
less questions than I had anticipated. Good or bad, I'm not sure.
Anyways, I've added a new section on case-classes and I've revised
the section on Try. It wasn't really driving home the point for
why it is useful and/or wonderful.